Try out a very non-noisy dependabot
Setting it at a monthly interval would be interesting to try out. The "daily" we have in standard/standard and others is a bit too noisy for my taste. "weekly" might be better, but I would like to try "monthly" first and increase it if we find it to be too infrequent.